+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Added a check that dimensions of any `inits` given match the calculated dimensions from the model formulae.
 
+* Added a check that if multiple repeated longitudinal outcoems were given, that each subject contributes at least one measurement per outcome.
+
 ## Bug patches
 
 * Patched a major bug in `gammaUpdate()` where ties in failure times were not being properly handled. The code for `gammaUpdate_approx()` always worked fine, as it was based only on the score vector. This bug manifested when `bootSE()` was called due to the resampling with replacement yielding datasets with many more ties than in the original dataset used to fit the model. To fix it, the information matrix required scaling at each failure time by the number of failures in the data. The formula for the information matrix in the Technical Details vignette has also been updated.